pyproject.toml
[tool.poetry]
name = "bookseer"
version = "0.1.0"
description = ""
authors = ["VladGalchevsky"]

[tool.poetry.dependencies]
python = "3.10.6"
Django = "3.2"
Pillow = "9.3.0"
psycopg2-binary = "2.9.5"
django-crispy-forms = "1.14.0"
django-colorfield = "^0.8.0"
django-colorful = "^1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
